[refactor] move ModalTemplate location
- move from container forder to template folder
