+++ src/components/templates/ModalTemplate/styled.ts
